bug 1498533, add serialization support, r=stas
Summary:
This adds an option to merge_channels to keep the oldest
instead of the newest variant, and uses that to serialize a file.

The file structure is taken from the en-US template.

This only supports Fluent, Android's string.xml and .properties.
Other file formats supported by compare-locales might work,
but will be added later (and get tests added).

This comes as two changesets, the first of which is just
technical and doesn't really require review.
It's requiring a follow-up change in cross-channel, so I
don't want to just land that.

Reviewers: stas

Tags: #secure-revision

Bug #: 1498533

Differential Revision: https://phabricator.services.mozilla.com/D8975