Bug 1557217, create mochibake on encoding errors, and warn. r=flod
There's a bit of a refactor in the base checker, which is now
good to use standalone. I use it for defines and include files, now.
That's because previously, the encoding errors would have been
reported. Sadly, only in stdout, and not in any structured data.

I've went for a warning instead of an error for two reasons:

- We might want to have an actual mochibake in localized content
- We shall run linters with fatal warnings in most en-US
  scenarios, which is as good as an error.

I doubt this will ever really be triggered by localizers anymore,
as they're basically all in Pontoon, where they can't screw up
encodings.

Differential Revision: https://phabricator.services.mozilla.com/D33987