Start on typing, via automation
Use monkeytype and retype to generate a first pass at typing the project,
using the test suite.

To recreate, first add a script monkeyrun.py:

    import pytest

    pytest.main(["--pyargs", "tests"])

Then run it to trace how functions are called:

    pip install monkeytype
    monkeytype run monkeyrun.py

After this, get the list of typable modules:

    monkeytype list-modules

For most, types can be directly applied:

    monkeytype apply --pep_563 l10n_parser.FOO

But because it's a bit buggy for inner classes, some need to have stub files generated:

    mkdir types
    monkeytype stub l10n_parser.FOO > types/FOO.pyi

Then, manually fix the .pyi file and use retype to apply it to the source:

    pip install retype
    retype l10n_parser/FOO.py
    mv typed-src/FOO.py l10n_parser/FOO.py

The results of this may need some manual cleanup,
for both the affected file as well as its dependents.